HAPP-1332: Create FetchCovidOrdersUseCase
